tweak printf format strlen return value

+++ src/format_output.c
@@ -866,7 +866,7 @@ format_oligos(FILE *f,
     }
   }
   if(print_primers == 1) { 
-    fprintf(f, "SEQUENCE SIZE: %d\n", strlen(sa->sequence));
+    fprintf(f, "SEQUENCE SIZE: %ld\n", (long int) strlen(sa->sequence));
     fprintf(f, "INCLUDED REGION SIZE: %d\n\n", sa->incl_l);
     print_pair_array(f, "TARGETS", sa->tar2.count, sa->tar2.pairs, pa, sa);
     print_pair_array(f, "EXCLUDED REGIONS", sa->excl2.count, sa->excl2.pairs, pa, sa);
